More tests. Refactor common methods into separate files
